requirements.txt
# =========================================================================== #
#                         CORE DEEP LEARNING                                  #
# =========================================================================== #
torch>=2.0
torchvision>=0.15

# =========================================================================== #
#                         DATA & MATHEMATICS                                  #
# =========================================================================== #
numpy>=1.23
pandas>=1.5
scikit-learn>=1.2
Pillow>=9.4.0

# =========================================================================== #
#                    CONFIGURATION & VALIDATION                               #
# =========================================================================== #
pydantic>=2.0
pyyaml>=6.0

# =========================================================================== #
#                    HYPERPARAMETER OPTIMIZATION                              #
# =========================================================================== #
optuna>=3.5.0
plotly>=6.5.2

# =========================================================================== #
#                              TESTING                                        #
# =========================================================================== #
pytest>=7.0

# =========================================================================== #
#                    VISUALIZATION & REPORTING                                #
# =========================================================================== #
matplotlib>=3.6
xlsxwriter>=3.0
openpyxl>=3.1

# =========================================================================== #
#                             UTILITIES                                       #
# =========================================================================== #
tqdm>=4.64
requests>=2.28
psutil>=5.9